Reject siblings with no content type
When a deletion conflict occurs, attempting to read the `data` property
of an `RObject` results in an exception being generated. The exception is
raised because `riak-client` is unable to find a serializer for an object
that is marked for deletion (read: no content_type). Since we cannot directly
access `data`, inspect all of the siblings and reject the items having a `nil`
`content_type`. This allows for conflict resolution to continue for the remaining
siblings.